Finished VehicleAccidentReport
Adjusted all the fields in the view to match the model and PHPMyAdmin.

Also changed fields inside the vehicleAccidentAppModel so that it matches the PHPMyAdmin

Lastly, Changed few typing errors so that the fields would match.

Name_of_the_person_Completing_the_report still says it is duplicated inside the model but could not find duplication, even using CTRL+F